Add files via upload
These files are in a pre-release state.  This means that parts are subject to change, which may make old versions incompatible with future updates.  Also, the sample code and help file are preliminary, since the code needs to be closer to done before they should be worked on.  All code is "use at your own risk" at this point.